requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
_tflow_select
absl-py
astor
blas
ca-certificates
certifi
gast
grpcio
h5py
hdf5
icc_rt
intel-openmp
keras-applications
keras-preprocessing
libprotobuf
markdown
mkl
mkl_fft
mkl_random
numpy
numpy-base
openssl
pandas
pip
protobuf
pyreadline
python
python-dateutil
pytz
scipy
setuptools
six
sqlite
tensorboard
tensorflow
tensorflow-base
termcolor
vc
vs2015_runtime
werkzeug
wheel
wincertstore
zlib